backends/requirements.txt
asgiref==3.4.1
Django==3.2.8
django-environ==0.8.1
djangorestframework==3.12.4
Pillow==8.4.0
psycopg2-binary==2.9.1
django-filter==21.1
drf-yasg==1.20.0
django-cors-headers==3.10.0
gunicorn==20.1.0
Pillow==8.4.0
pyexcel_odsr==0.6.0
drf-spectacular==0.24.2
google-api-python-client==2.66.0